chore: split server/coreSystem.ts into multiple files
